Bug fixed

+++ scripts/copy_baseline_IRs.sh
@@ -66,12 +66,33 @@ if ! test -z ${NOELLE_SPEC} ; then
 fi
 
 # Cache the bitcode files
+echo "Copy the IRs" ;
 outputDir="${origDir}/results/current_machine" ;
 for i in `ls */benchmarks/*/*.bc` ; do
-  echo $i ;
+  fileName="`basename $i`";
+  if [[ $fileName == "NOELLE"* ]] ; then
+    continue ;
+  fi
+  if [[ $fileName == "all.bc" ]] ; then
+    continue ;
+  fi
 
+  # Create the output directory
   dirName="`dirname $i`" ;
-  echo $dirName
   mkdir -p ${outputDir}/IR/${dirName} ;
-  cp $i ${outputDir}/IR/${dirName}/NOELLE_input.bc ;
+
+  # Copy the IR
+  if test -f ${outputDir}/IR/${dirName}/NOELLE_input.bc ; then
+    cmp $i ${outputDir}/IR/${dirName}/NOELLE_input.bc &> /dev/null ;
+    if test $? -ne 0 ; then
+      echo "ERROR: The IR in \"${outputDir}\" does not match the single bitcode IR stored in the benchmark suite" ;
+      echo "       Previous IR = ${outputDir}/IR/${dirName}/NOELLE_input.bc";
+      echo "       IR in the benchmark suite: $i" ;
+      exit 1;
+    fi
+
+  else 
+    echo "  Copy the IR $i" ;
+    cp $i ${outputDir}/IR/${dirName}/NOELLE_input.bc ;
+  fi
 done
